backend/requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14,<3.15
djoser>=2.2,<2.3
djangorestframework-simplejwt>=5.3,<5.4
psycopg2-binary>=2.9,<2.10
Pillow>=10.0,<10.1
django-filter>=23.2,<23.3